requirements.txt
asgiref==3.3.4
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
djangorestframework-simplejwt==4.7.1
Pillow==8.2.0

astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
django-crispy-forms==1.11.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
requests==2.22.0
six==1.13.0
sslcommerz-python==0.0.7
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
